test(update): drive the real `fresh` binary instead of a stand-in
The spine test drove a purpose-built harness that called the engine directly.
That proved the engine worked and could not prove that `fresh` did: flag
parsing, the endpoint overrides, the feature gate and the exit-code mapping the
editor's update indicator keys off were all untested, and a stand-in that
duplicates the caller is a second implementation free to drift from the one
that ships.

It now spawns CARGO_BIN_EXE_fresh and drives it as a user would:
`fresh --cmd update --yes --releases-url U --download-base U`. No test-only
surface was needed — those two flags already ship for air-gapped and enterprise
mirrors, and setting them marks the endpoints untrusted, which is what keeps
the test away from sudo and skips an attestation check a local server could
never satisfy. The version is faked purely server-side: the fabricated feed
announces 99.9.9, so the binary's own CURRENT_VERSION is genuinely older and
the upgrade path is entered for the real reason. No second build, and nothing
that could leak into a release.

Six cases: the swap lands and stays runnable; --check reports and changes
nothing; without --yes it describes the swap and exits ActionRequired; a bad
checksum aborts leaving the binary byte-identical; a missing asset fails
without touching it; an install with no receipt refuses to swap at all.

Verified by mutation that this reaches code the harness could not: collapsing
the ActionRequired exit mapping in main.rs to Ok(()) — a pure CLI-layer defect
— fails two of the six.

The harness binary is deleted, which also removes a test-only bin from a
published library crate. What stays in fresh-update is the part that needs no
binary: tests/release_contract.rs, asserting the engine asks for the extension
the workflows actually publish.
